Send V2 for Matlab Scripts
This push includes functions to read/process datasets on the user's system, log scale the data, and also makes edits/updates to the old class and model functions which allow for "stacking" multiple approach velocities. The code is also general enough to work for only a single dataset---the user need only pass a cell array with one element to each of the force, indentation, time, and tipSize arguments. For the case with multiple datasets fit simultaneously, the user would provide a cell array of each where every cell array entry would correspond to another experiment/force curve.