Added Thin-Sample Correction for Spherical and Conical Indentation
Implemented the thin-sample correction as dictated by Garcia et. al (2020 & 2018). This code remains untested, and VERY importantly, the coefficients for the generalized voigt model MUST BE VERIFIED! The coefficients should be agnostic of the model chosen, but we do not know if they change form when the model is inverted (i.e. when we solve for h as the output instead of as the input, as Garcia did).