Sending functions to read IBW and modifications to example.m
Modified example.m to run in a batch format---i.e., if the directory chosen contains subdirectories, it will loop through those for analysis. Otherwise it just performs the analysis on that directory. This is to make analyzing the results for the paper easier. Also added comments and fixed a couple bugs in LoadAFMData.m.