Expanded Functionality for JPK Maps
Added all the code necessary to load JPK QI Map Files. It should also work for JPK Force Files. It remains to test the functionality for maps that contain all the information we need for fitting. There may be some issues with the averaging, but that remains to be debugged once we get more (complete) data.